# Only enforce cardinality limits when the attribute set does not already exist