--- a/sdk/metric/internal/aggregate/exponential_histogram.go
+++ b/sdk/metric/internal/aggregate/exponential_histogram.go
@@ -338,18 +338,17 @@ func (e *expoHistogram[N]) measure(
 	e.valuesMu.Lock()
 	defer e.valuesMu.Unlock()
 
-	attr := fltrAttr
-	v, ok := e.values[attr.Equivalent()]
+	v, ok := e.values[fltrAttr.Equivalent()]
 	if !ok {
-		attr = e.limit.Attributes(fltrAttr, e.values)
+		fltrAttr = e.limit.Attributes(fltrAttr, e.values)
 		// if we overflowed, make sure we add to the existing overflow series
 		// if it already exists.
-		v, ok = e.values[attr.Equivalent()]
+		v, ok = e.values[fltrAttr.Equivalent()]
 		if !ok {
-			v = newExpoHistogramDataPoint[N](attr, e.maxSize, e.maxScale, e.noMinMax, e.noSum)
-			v.res = e.newRes(attr)
+			v = newExpoHistogramDataPoint[N](fltrAttr, e.maxSize, e.maxScale, e.noMinMax, e.noSum)
+			v.res = e.newRes(fltrAttr)
 
-			e.values[attr.Equivalent()] = v
+			e.values[fltrAttr.Equivalent()] = v
 		}
 	}
 	v.record(value)